% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotypepredictionforallinteractions.R
\name{ERRBLUP_Stepwise}
\alias{ERRBLUP_Stepwise}
\title{ERRBLUP Phenotype Prediction Function Relying On The Out Put Of ERRBLUP Relationship Matrix Function}
\usage{
ERRBLUP_Stepwise(Pheno, G_ERRBLUP)
}
\arguments{
\item{Pheno}{A numeric vector of phenotypes}

\item{G_ERRBLUP}{ERRBLUP relationship matrix}
}
\value{
A numeric vector of both phenotype estimations of training set and phenotype predictions of test set based on ERRBLUP method
}
\description{
Function to do phenotype prediction based on all pairwise SNP interactions
}
\examples{
library(BGLR)
data(wheat)
m <- Recodemarkers(wheat.X)
G_ERRBLUP <- Gall(m, cores=15)
G <- G_ERRBLUP$G
N <- length(Phenotype)
n <- 60
test <- sample(1:N,n)
pheno <- Phenotype
pheno[test] <- NA
Pred_ERRBLUP <- ERRBLUP_Stepwise(pheno, G)

}
