% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allpairwiseinteractionsrelationshipmatrix.R
\name{Gall}
\alias{Gall}
\title{ERRBLUP Relationship Matrix Function}
\usage{
Gall(m)
}
\arguments{
\item{m}{{0,1,2} or {0,2} coded marker matrix with individuals in the rows and the markers in the columns}
}
\value{
ERRBLUP relationship matrix with row names and column names of all the individuals
}
\description{
Function to generate relationship matrix based on all pairwise SNP interactions
}
\examples{
library(BGLR)
data(wheat)
m <- Recodemarkers(wheat.X[1:100,])
rownames(m) <- names(wheat.Y[1:100,3])
G_ERRBLUP <- Gall(m)

}
