% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Topinteractionsbasedonhighesteffectrelationshipmatrix.R
\name{Gtop}
\alias{Gtop}
\title{sERRBLUP Relationship Matrix Function}
\usage{
Gtop(m, Estimations, k, cores = 1)
}
\arguments{
\item{m}{The \code{{0,1,2}} or \code{{0,2}} coded marker matrix with individuals in the rows and the markers in the columns}

\item{Estimations}{A numeric vector of all estimated pairwise SNP interaction effect sizes or all estimated pairwise SNP interaction effect variances}

\item{k}{A desired proportion of SNP interactions to be included in the model}

\item{cores}{The number of cores with the default value of \code{1}}
}
\value{
sERRBLUP Relationship matrix when \code{k} percent of pairwise SNP interactions maintained in the model
}
\description{
Function to generate relationship matrix based on the desired proportion of pairwise SNP interactions
}
\examples{
library(BGLR)
data(wheat)
N <- length(Phenotype)
n <- 60
test <- sample(1:N,n)
pheno <- Phenotype
pheno[test] <- NA
m <- Recodemarkers(wheat.X)
G_ERRBLUP <- Gall(m, cores=15)
G <- G_ERRBLUP$G
P <- G_ERRBLUP$P
Estimation <- SNP_Effect_Var(m, pheno, G, P, cores=15)
t_hat <- Estimation$Effect
sigma_hat <- Estimation$Effect.Var
k <- 10
Gtop_effect <- Gtop(m, t_hat, k, cores=15)
Gtop_var <- Gtop(m, sigma_hat, k, cores=15)

}
