% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Phenotype}
\alias{Phenotype}
\title{Simulated Wheat Phenotypes}
\format{
A Phenotypic vector of length \code{599}
}
\usage{
Phenotype
}
\description{
Simulated phenotypes from Wheat dataset ("BGLR" package) genotype by randomly selection of \code{80} SNPs from marker matrix and generation of the phenotypic trait based on all \code{4} possible genotype combinations  \code{{00, 02, 20, 22}} of \code{{0,2}} coded marlers resulting in purely epistatic trait with the heritability of \code{1}
}
\keyword{datasets}
