% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recodingthemarkermatrix.R
\name{Recodemarkers}
\alias{Recodemarkers}
\title{Marker Recoding Function}
\usage{
Recodemarkers(M)
}
\arguments{
\item{M}{The original marker matrix of \code{{-1,0,1}} or \code{{0,1}} or character coded markers}
}
\value{
A \code{{0,1,2}} or \code{{0,2}} coded marker matrix with reporting how the recoding was done in the original marker matrix
}
\description{
Function to recode marker matrix to \code{{0,2}} or \code{{0,1,2}} coded marker matrix
}
\examples{
library(BGLR)
data(wheat)
m <- Recodemarkers(wheat.X)

}
