% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotypepredictionfortopinteractions.R
\name{sERRBLUP}
\alias{sERRBLUP}
\title{sERRBLUP Phenotype Prediction Function}
\usage{
sERRBLUP(Pheno_train, Gtop, Trainset)
}
\arguments{
\item{Pheno_train}{A subset of one numeric phenotype vector as a training set with names for each phenotypic value}

\item{Gtop}{sERRBLUP Relationship matrix for the k percent of pairwise SNP interactions with row names and column names of all the individuals}

\item{Trainset}{A vector of individuals which are in the training set}
}
\value{
A numeric vector of both phenotype estimations of training set and phenotype predictions of test set based on sERRBLUP method
}
\description{
Function to do phenotype prediction based on the desired proportion of pairwise SNP interactions
}
\examples{
library(BGLR)
data(wheat)
geno <- wheat.X
t1 <- sample(1:ncol(geno), 20)
t2 <- sample(1:ncol(geno), 20)
y1 <- rowSums((geno[,t1]==2) * (geno[,t2]==2))
t1 <- sample(1:ncol(geno), 20)
t2 <- sample(1:ncol(geno), 20)
y2 <- rowSums((geno[,t1]==2) * (geno[,t2]==0))
t1 <- sample(1:ncol(geno), 20)
t2 <- sample(1:ncol(geno), 20)
y3 <- rowSums((geno[,t1]==0) * (geno[,t2]==2))
t1 <- sample(1:ncol(geno), 20)
t2 <- sample(1:ncol(geno), 20)
y4 <- rowSums((geno[,t1]==0) * (geno[,t2]==0))
y <- y1+y2+y3+y4
pheno <- scale(y)
names(pheno) <- names(wheat.Y[,1])
N <- length(pheno)
n <- 60
test <- sample(1:N,n)
training <- (1:N)[-test]
pheno_train <- pheno[training]
m <- Recodemarkers(wheat.X)
rownames(m) <- names(pheno)
G_ERRBLUP <- Gall(m, cores=15)
G <- G_ERRBLUP$G
pi <- G_ERRBLUP$Pi
Estimation <- SNP_effect_var(m, pheno_train, G, pi, training, cores=15)
t_hat <- Estimation$effect
sigma_hat <- Estimation$effectvar
k <- 10
Gtop_effect <- Gtop(m, t_hat, k, cores=15)
Gtop_var <- Gtop(m, sigma_hat, k, cores=15)
sERRBLUP_effect <- sERRBLUP(pheno_train, Gtop_effect, training)
sERRBLUP_var <- sERRBLUP(pheno_train, Gtop_var, training)

}
