% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sERRBLUP_bivar.R
\name{sERRBLUP_BV}
\alias{sERRBLUP_BV}
\title{Bivariate sERRBLUP Phenotype Prediction Function}
\usage{
sERRBLUP_BV(M, y1, y2, k, iters = 20, tolparinv = 1e-06, cores = 1)
}
\arguments{
\item{M}{The original marker matrix of \code{{-1,0,1}}, \code{{0,1}}, \code{{0,1,2}}, \code{{0,2}} or character coded markers with named individuals in the rows and the markers in the columns}

\item{y1}{A numeric vector of named phenotypes for the target triat to be predicted in bivariate sERRBLUP}

\item{y2}{A numeric vector of named phenotypes for the additional trait to be estimated in bivariate sERRBLUP}

\item{k}{A desired proportion of SNP interactions to be included in the model}

\item{iters}{Maximum number of iterations allowed with the default value of \code{20} (The parameter of mmer function from sommer package)}

\item{tolparinv}{Tolerance parameter for matrix inverse used when singularities are encountered in the estimation procedure with the default value of \code{1e-06} (The parameter of mmer function from sommer package)}

\item{cores}{The number of cores with the default value of \code{1}}
}
\value{
A list of eight components:

\describe{
  \item{Recodedmarkers}{A \code{{0,1,2}} or \code{{0,2}} coded marker matrix}
  \item{Relationshipmatrix}{A list of two components: ERRBLUP relationship matrix (G) and a vector of all genotype combinations frequencies in the population (P)}
  \item{Effect.Relationshipmatrix}{sERRBLUP Relationship matrix when \code{k} percent of pairwise SNP interactions maintained in the model based on absolute effect size estimations}
  \item{Var.Relationshipmatrix}{sERRBLUP Relationship matrix when \code{k} percent of pairwise SNP interactions maintained in the model based on effect variance estimations}
  \item{Effect}{A numeric vector of all estimated pairwise SNP interaction effect sizes}
  \item{Effect.Variance}{A numeric vector of all estimated pairwise SNP interaction effect variances}
  \item{Effect.Predictions}{A dataframe of both phenotype predictions of target triat test set, phenotype estimations of target triat training set and phenotype estimations of additional trait full set in bivarite sERRBLUP method based on effect sizes}
  \item{Var.Prediction}{A dataframe of both phenotype predictions of target triat test set, phenotype estimations of target triat training set and phenotype estimations of additional trait full set in bivarite sERRBLUP method based on effect variances}
}
}
\description{
Function to do phenotype prediction based on desired proportion of pairwise SNP interactions in bivariate model
}
\examples{
library(BGLR)
data(wheat)
y1 <- Phenotype_BV[,1]
N <- length(y1)
n <- 60
test <- sample(1:N,n)
y1[test] <- NA
y2 <- Phenotype_BV[,2]
M <- wheat.X
rownames(M) <- names(y1)
sERRBLUP_Bivar <- sERRBLUP_BV(M, y1, y2, 10, iters=20, tolparinv= 1e-06, cores=15)

}
