% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sERRBLUP_Bivar_Stepwise.R
\name{sERRBLUP_BV_Stepwise}
\alias{sERRBLUP_BV_Stepwise}
\title{Bivariate sERRBLUP Phenotype Prediction Function Relying On The Out Put Of sERRBLUP Relationship Matrix Function}
\usage{
sERRBLUP_BV_Stepwise(y1, y2, Gtop, iters = 20, tolparinv = 1e-06)
}
\arguments{
\item{y1}{A numeric vector of named phenotypes for the target triat to be predicted in bivariate sERRBLUP}

\item{y2}{A numeric vector of named phenotypes for the additional trait to be estimated in bivariate sERRBLUP}

\item{Gtop}{sERRBLUP Relationship matrix for the \code{k} percent of pairwise SNP interactions with the row names and column names of all individuals}

\item{iters}{Maximum number of iterations allowed with the default value of \code{20} (The parameter of mmer function from sommer package)}

\item{tolparinv}{Tolerance parameter for matrix inverse used when singularities are encountered in the estimation procedure with the default value of \code{1e-06} (The parameter of mmer function from sommer package)}
}
\value{
A dataframe of both phenotype predictions of target triat test set, phenotype estimations of target triat training set and phenotype estimations of additional trait full set in bivarite sERRBLUP method
}
\description{
Function to do phenotype prediction based on the desired proportion of pairwise SNP interactions in bivariate model
}
\examples{
library(BGLR)
data(wheat)
y1 <- Phenotype_BV[,1]
N <- length(y1)
n <- 60
test <- sample(1:N,n)
y1[test] <- NA
y2 <- Phenotype_BV[,2]
m <- Recodemarkers(wheat.X)
G_ERRBLUP <- Gall(m, cores=15)
G <- G_ERRBLUP$G
P <- G_ERRBLUP$P
rownames(G) <- names(y1)
colnames(G) <- names(y1)
Estimation <- SNP_Effect_Var(m, y2, G, P, cores=15)
t_hat <- Estimation$Effect
sigma_hat <- Estimation$Effect.Var
k <- 10
Gtop_effect <- Gtop(m, t_hat, k, cores=15)
rownames(Gtop_effect) <- names(y1)
colnames(Gtop_effect) <- names(y1)
Gtop_var <- Gtop(m, sigma_hat, k, cores=15)
rownames(Gtop_var) <- names(y1)
colnames(Gtop_var) <- names(y1)
sERRBLUP_effect <- sERRBLUP_BV_Stepwise(y1 , y2, Gtop_effect, iters=20, tolparinv= 1e-06)
sERRBLUP_var <- sERRBLUP_BV_Stepwise(y1 , y2, Gtop_var, iters=20, tolparinv= 1e-06)

}
