% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfoldcorossvalidation_bivar.R
\name{sERRBLUP_BV_Test}
\alias{sERRBLUP_BV_Test}
\title{Bivariate sERRBLUP Predictive Ability Test}
\usage{
sERRBLUP_BV_Test(M, y1, y2, k, iters = 20, tolparinv = 1e-06, cores = 1)
}
\arguments{
\item{M}{The original marker matrix of \code{{-1,0,1}}, \code{{0,1}}, \code{{0,1,2}}, \code{{0,2}} or character coded markers with named individuals in the rows and the markers in the columns}

\item{y1}{A numeric vector of named phenotypes for the target triat to be predicted in bivariate sERRBLUP}

\item{y2}{A numeric vector of named phenotypes for the additional trait to be estimated in bivariate sERRBLUP}

\item{k}{A vector of desired proportions of SNP interactions to be included in the model wich is proposed to be \code{(10, 5, 1, 0.1)}}

\item{iters}{Maximum number of iterations allowed with the default value of \code{20} (The parameter of mmer function from sommer package)}

\item{tolparinv}{Tolerance parameter for matrix inverse used when singularities are encountered in the estimation procedure with the default value of \code{1e-06} (The parameter of mmer function from sommer package)}

\item{cores}{The number of cores with the default value of \code{1}}
}
\value{
A data frame of 3 components:

\describe{
  \item{Desired.Proportion}{The proportion of SNP interactions maintained in the bivariate sERRBLUP model}
  \item{PA.Effcet}{Bivariate sERRBLUP predictive ability based on effect sizes}
  \item{PA.Var}{Bivariate sERRBLUP predictive ability based on effect variances}
}
}
\description{
Function to provide predictive abilities of bivariate sERRBLUP model for series of diffetent proportions of pairwise SNP interactions to find the optimum proportion of interactions leading to the highest predictive ability
}
\examples{
library(BGLR)
data(wheat)
y1 <- Phenotype_BV[,1]
y2 <- Phenotype_BV[,2]
M <- wheat.X
rownames(M) <- names(y1)
K=c(10, 5, 1, 0.1)
Test <- sERRBLUP_BV_Test(M, y1, y2, K, iters=20, tolparinv= 1e-06, cores=15)

}
