% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfoldcorossvalidation.R
\name{sERRBLUP_Proportions_Test}
\alias{sERRBLUP_Proportions_Test}
\title{sERRBLUP Predictive Ability Test}
\usage{
sERRBLUP_Proportions_Test(M, Pheno, k, cores = 1)
}
\arguments{
\item{M}{The original marker matrix of \code{{-1,0,1}}, \code{{0,1}}, \code{{0,1,2}}, \code{{0,2}} or character coded markers}

\item{Pheno}{A numeric vector of phenotypes}

\item{k}{A vector of desired proportions of SNP interactions to be included in the model wich is proposed to be \code{(10, 5, 1, 0.1)}}

\item{cores}{The number of cores with the default value of \code{1}}
}
\value{
A data frame of 3 components:

\describe{
  \item{Desired.Proportion}{The proportion of SNP interactions maintained in the sERRBLUP model}
  \item{PA.Effcet}{sERRBLUP predictive ability based on effect sizes}
  \item{PA.Var}{sERRBLUP predictive ability based on effect variances}
}
}
\description{
Function to provide predictive abilities of sERRBLUP model for series of diffetent proportions of pairwise SNP interactions to find the optimum proportion of interactions leading to the highest predictive ability
}
\examples{
library(BGLR)
data(wheat)
N <- length(Phenotype)
n <- 60
test <- sample(1:N,n)
M <- wheat.X
pheno <- Phenotype
pheno[test] <- NA
k=c(10, 5, 1, 0.1)
cross_val <- sERRBLUP_Proportions_Test(M, pheno, k, cores=15)

}
