% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllSNPeffectcalculation.R
\name{SNP_effect_var}
\alias{SNP_effect_var}
\title{Pairwise SNP Interaction Effects and variances Estimation Function}
\usage{
SNP_effect_var(m, Pheno, G_ERRBLUP, P, cores = 1)
}
\arguments{
\item{m}{{0,1,2} or {0,2} coded marker matrix with individuals in the rows and the markers in the columns}

\item{Pheno}{A numeric vector of phenotypes}

\item{G_ERRBLUP}{EERRBLUP relationship matrix}

\item{P}{A vector of all genotype combinations frequencies in the population}

\item{cores}{The number of cores with the default value of 1}
}
\value{
A list of two components effect and effectvar

\describe{
  \item{Effect}{A numeric vector of all estimated pairwise SNP interaction effects}
  \item{Effect.Var}{A numeric vector of all estimated pairwise SNP interaction effects variances}
}
}
\description{
Function to calculate all pairwise SNP interaction effects
}
\examples{
library(BGLR)
data(wheat)
N <- length(Phenotype)
n <- 60
test <- sample(1:N,n)
Phenotype[test] <- NA
m <- Recodemarkers(wheat.X)
G_ERRBLUP <- Gall(m, cores=15)
G <- G_ERRBLUP$G
P <- G_ERRBLUP$P
Estimation <- SNP_effect_var(m, Phenotype, G, P, cores=15)
t_hat <- Estimation$Effect
sigma_hat <- Estimation$Effect.Var

}
