% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllSNPeffectvariancecalculation.R
\name{SNP_var}
\alias{SNP_var}
\title{Pairwise SNP Interaction Effects Variances Function}
\usage{
SNP_var(m, pheno_train, t_hat)
}
\arguments{
\item{m}{{0,1,2} or {0,2} coding marker matrix with individuals in the rows and the markers in the columns}

\item{pheno_train}{A subset of one numeric phenotype vector as a training set with names for each phenotypic value}

\item{t_hat}{A numeric vector of all estimated pairwise SNP interaction effects}
}
\value{
A numeric vector of all pairwise SNP interaction effects variances
}
\description{
Function to calculate all pairwise SNP interaction effects varinacs
}
\examples{
library(BGLR)
data(wheat)
pheno <- wheat.Y[1:100,1]
pheno_train <- pheno[1:round(4*length(pheno)/5)]
m <- Recodemarker(wheat.X[1:100,])
rownames(m) <- names(pheno)
G_ERRBLUP <- Gall(m)
t_hat <- SNP_effect(m, pheno_train, G_ERRBLUP)
sigma_hat <- SNP_var(m, pheno_train, t_hat)

}
