% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERRBLUP.R
\name{ERRBLUP}
\alias{ERRBLUP}
\title{ERRBLUP Phenotype Prediction Function}
\usage{
ERRBLUP(M, Pheno, cores = 1)
}
\arguments{
\item{M}{A {-1, 0, 1} or {0, 1} or character coded marker matrix}

\item{Pheno}{A numeric vector of phenotypes}

\item{cores}{The number of cores with the default value of 1}
}
\value{
A list of three components:

\describe{
  \item{Recodedmarkers}{A {0, 1, 2} or {0, 2} coded marker matrix}
  \item{Relationshipmatrix}{A list of two components: ERRBLUP relationship matrix (G) and A vector of all genotype combinations frequencies in the population}
  \item{Predictions}{A numeric vector of both phenotype estimations of training set and phenotype predictions of test set based on ERRBLUP method}
}
}
\description{
Function to do phenotype prediction based on all pairwise SNP interactions
}
\examples{
library(BGLR)
data(wheat)
M <- wheat.X[,1:10]
N <- length(Phenotype)
n <- 60
test <- sample(1:N,n)
Phenotype[test] <- NA
ERRBLUP <- ERRBLUP(M, Phenotype)

}
