% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERRBLUP.R
\name{ERRBLUP}
\alias{ERRBLUP}
\title{ERRBLUP Phenotype Prediction Function}
\usage{
ERRBLUP(M, Pheno, cores = 1)
}
\arguments{
\item{M}{The original marker matrix of \code{{-1,0,1}}, \code{{0,1}}, \code{{0,1,2}}, \code{{0,2}} or character coded markers}

\item{Pheno}{A numeric vector of phenotypes}

\item{cores}{The number of cores with the default value of \code{1}}
}
\value{
A list of three components:

\describe{
  \item{Recodedmarkers}{A \code{{0,1,2}} or \code{{0,2}} coded marker matrix}
  \item{Relationshipmatrix}{A list of two components: ERRBLUP relationship matrix (G) and a vector of all genotype combinations frequencies in the population(P)}
  \item{Predictions}{A numeric vector of both phenotype estimations of training set and phenotype predictions of test set based on ERRBLUP method}
}
}
\description{
Function to do phenotype prediction based on all pairwise SNP interactions
}
\examples{
library(BGLR)
data(wheat)
M <- wheat.X
N <- length(Phenotype)
n <- 60
test <- sample(1:N,n)
pheno <- Phenotype
pheno[test] <- NA
ERRBLUP <- ERRBLUP(M, pheno)

}
