% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allpairwiseinteractionsrelationshipmatrix.R
\name{Gall}
\alias{Gall}
\title{ERRBLUP Relationship Matrix Function}
\usage{
Gall(m, cores = 1)
}
\arguments{
\item{m}{{0,1,2} or {0,2} coded marker matrix with individuals in the rows and the markers in the columns}

\item{cores}{The number of cores with the default value of 1}
}
\value{
A list of two components G and Pi

\describe{
  \item{G}{ERRBLUP relationship matrix with row names and column names of all the individuals}
  \item{Pi}{A vector of all genotype combinations frequencies in the population}
}
}
\description{
Function to generate relationship matrix based on all pairwise SNP interactions
}
\examples{
library(BGLR)
data(wheat)
m <- Recodemarkers(wheat.X)
rownames(m) <- names(wheat.Y[,3])
G_ERRBLUP <- Gall(m, cores=15)

}
