% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Topinteractionsbasedonhighesteffectrelationshipmatrix.R
\name{Gtop}
\alias{Gtop}
\title{sERRBLUP Relationship Matrix Function based on effect sizes selection}
\usage{
Gtop(M, pheno, estimations, k)
}
\arguments{
\item{M}{{0,1,2} or {0,2} coded Marker matrix with individuals in the rows and the markers in the columns}

\item{pheno}{A phenotype vector with names for each phenotypic value}

\item{estimations}{A vector of all estimated pairwise SNP interaction effects or all estimated variances pairwise SNP interaction effects variances}

\item{k}{Desired percentage of SNP interactions to be included in the model}
}
\value{
sERRBLUP Relationship matrix for the k percent of pairwise SNP interactions with row names and column names of all the individuals
}
\description{
Function to generate relationship matrix based on the top desired estimated pairwise SNP interaction effects
}
\examples{
library(BGLR)
data(wheat)
pheno <- wheat.Y[1:100,1]
pheno_train <- pheno[1:round(4*length(pheno)/5)]
M <- Recodemarker(wheat.X[1:100,])
rownames(M) <- names(pheno)
G_all <- Gall(M)
u_hat <- SNP_effect(M, pheno_train, G_all)
u_hat_var <- SNP_var(M, pheno_train, u_hat)
k <- 10
Gtop_effect <- Gtop(M, pheno , u_hat, k)
Gtop_var <- Gtop(M, pheno , u_hat_var, k)

}
