% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Topinteractionsbasedonhighesteffectrelationshipmatrix.R
\name{Gtop}
\alias{Gtop}
\title{sERRBLUP Relationship Matrix Function}
\usage{
Gtop(m, pheno, estimations, k)
}
\arguments{
\item{m}{{0,1,2} or {0,2} coded marker matrix with individuals in the rows and the markers in the columns}

\item{pheno}{A numeric phenotype vector with names for each phenotypic value}

\item{estimations}{A numeric vector of all estimated pairwise SNP interaction effects or all estimated pairwise SNP interaction effects variances}

\item{k}{Desired proportion of SNP interactions to be included in the model}
}
\value{
sERRBLUP Relationship matrix for the k percent of pairwise SNP interactions with row names and column names of all the individuals
}
\description{
Function to generate relationship matrix based on the desired proportion of pairwise SNP interactions
}
\examples{
library(BGLR)
data(wheat)
pheno <- wheat.Y[1:100,1]
pheno_train <- pheno[1:round(4*length(pheno)/5)]
m <- Recodemarker(wheat.X[1:100,])
rownames(m) <- names(pheno)
G_ERRBLUP <- Gall(m)
t_hat <- SNP_effect(m, pheno_train, G_ERRBLUP)
sigma_hat <- SNP_var(m, pheno_train, t_hat)
k <- 10
Gtop_effect <- Gtop(m, pheno , t_hat, k)
Gtop_var <- Gtop(m, pheno , sigma_hat, k)

}
