% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Topinteractionsbasedonhighesteffectrelationshipmatrix.R
\name{Gtop}
\alias{Gtop}
\title{sERRBLUP Relationship Matrix Function}
\usage{
Gtop(m, Estimations, k, cores = 1)
}
\arguments{
\item{m}{{0,1,2} or {0,2} coded marker matrix with named individuals in the rows and the markers in the columns}

\item{Estimations}{A numeric vector of all estimated pairwise SNP interaction effects or all estimated pairwise SNP interaction effects variances}

\item{k}{Desired proportion of SNP interactions to be included in the model}

\item{cores}{The number of cores with the default value of 1}
}
\value{
sERRBLUP Relationship matrix for the k percent of pairwise SNP interactions
}
\description{
Function to generate relationship matrix based on the desired proportion of pairwise SNP interactions
}
\examples{
library(BGLR)
data(wheat)
N <- length(Phenotype)
n <- 60
test <- sample(1:N,n)
Phenotype[test] <- NA
m <- Recodemarkers(wheat.X[,1:10])
G_ERRBLUP <- Gall(m, cores=15)
G <- G_ERRBLUP$G
P <- G_ERRBLUP$P
Estimation <- SNP_effect_var(m, Phenotype, G, P, cores=15)
t_hat <- Estimation$effect
sigma_hat <- Estimation$effectvar
k <- 10
Gtop_effect <- Gtop(m, t_hat, k, cores=15)
Gtop_var <- Gtop(m, sigma_hat, k, cores=15)

}
