% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Phenotype}
\alias{Phenotype}
\title{Simulated wheat Phenotypes}
\format{A Phenotypic vector of length 599}
\usage{
Phenotype
}
\description{
Simulated phenotypes from Wheat dataset ("BGLR" package) genotype by randomly selection of 80 SNPs from marker matrix and generation of the phenotypic trait based on all 4 possible genotype combinations of {0,2} coded marlers (00, 02, 20, 22) resulting in purely epistatic trait with the heritability of 1
}
\keyword{datasets}
