% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recodingthemarkermatrix.R
\name{Recodemarkers}
\alias{Recodemarkers}
\title{Marker Recoding Function}
\usage{
Recodemarkers(M)
}
\arguments{
\item{M}{A {-1, 0, 1} or {0, 1} coded marker matrix}
}
\value{
A {0, 1, 2} or {0, 2} coded marker matrix
}
\description{
Function to recode marker data from {-1, 0, 1} to {0, 1, 2} or from {0, 1} to {0, 2} marker coding
}
\examples{
library(BGLR)
data(wheat)
m <- Recodemarkers(wheat.X)

}
