% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recodingthemarkermatrix.R
\name{Recodemarkers}
\alias{Recodemarkers}
\title{Marker Recoding Function}
\usage{
Recodemarkers(M)
}
\arguments{
\item{M}{A {-1, 0, 1} or {0, 1} or character coded marker matrix}
}
\value{
A {0, 1, 2} or {0, 2} coded marker matrix. The function reports how the recoding was implemented in the input marker matrix.
}
\description{
Function to recode marker matrix to {0, 2} or {0, 1, 2} coded marker matrix
}
\examples{
library(BGLR)
data(wheat)
m <- Recodemarkers(wheat.X)

}
