% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotypepredictionfortopinteractions.R
\name{sERRBLUP}
\alias{sERRBLUP}
\title{sERRBLUP Phenotype Prediction Function}
\usage{
sERRBLUP(pheno_train, Gall, Gtop)
}
\arguments{
\item{pheno_train}{A subset of one phenotype vector as a training set with names for each phenotypic value}

\item{Gall}{ERRBLUP relationship matrix with row names and column names of all the individuals}

\item{Gtop}{sERRBLUP Relationship matrix for the top k percent of pairwise SNP interactions with row names and column names of all the individuals}
}
\value{
A vector of both phenotype estimations of training set and phenotype predictions of test set based on sERRBLUP method
}
\description{
Function to do phenotype prediction based on the top desired pairwise SNP interaction
}
\examples{
library(BGLR)
data(wheat)
pheno <- wheat.Y[,1]
pheno_train <- pheno[1:round(4*length(pheno)/5)]
M <- Recodemarker(wheat.X)
rownames(M) <- names(pheno)
G_all <- Gall(M)
u_hat <- SNP_effect(M, pheno_train, G_all)
u_hat_var <- SNP_var(M, pheno_train, u_hat)
k <- 10
Gtop_effect <- Gtop(M, pheno , u_hat, k)
Gtop_var <- Gtop(M, pheno , u_hat_var, k)
sERRBLUP_effect <- sERRBLUP(pheno_train, G_all, Gtop_effect)
sERRBLUP_var <- sERRBLUP(pheno_train, G_all, Gtop_var)

}
