% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sERRBLUP.R
\name{sERRBLUP}
\alias{sERRBLUP}
\title{sERRBLUP Phenotype Prediction Function}
\usage{
sERRBLUP(M, Pheno, k, cores = 1)
}
\arguments{
\item{M}{A {-1, 0, 1} or {0, 1} or character coded marker matrix}

\item{Pheno}{A numeric vector of phenotypes}

\item{k}{Desired proportion of SNP interactions to be included in the model}

\item{cores}{The number of cores with the default value of 1}
}
\value{
A list of eight components:

\describe{
  \item{Recodedmarkers}{A {0, 1, 2} or {0, 2} coded marker matrix}
  \item{Relationshipmatrix}{A list of two components: ERRBLUP relationship matrix (G) and A vector of all genotype combinations frequencies in the population}
  \item{Effect.Relationshipmatrix}{sERRBLUP Relationship matrix for the k percent of pairwise SNP interactions based on effect size estimations}
  \item{var.Relationshipmatrix}{sERRBLUP Relationship matrix for the k percent of pairwise SNP interactions baed on effect size variznce estimations}
  \item{Effect}{A numeric vector of all estimated pairwise SNP interaction effects}
  \item{Effect.variance}{A numeric vector of all estimated pairwise SNP interaction effects variances}
  \item{Effect.Predictions}{A numeric vector of both phenotype estimations of training set and phenotype predictions of test set based on effect sizes}
  \item{Var.prediction}{A numeric vector of both phenotype estimations of training set and phenotype predictions of test set based on effect sizes variances}
}
}
\description{
Function to do phenotype prediction based on all pairwise SNP interactions
}
\examples{
library(BGLR)
data(wheat)
N <- length(Phenotype)
n <- 60
test <- sample(1:N,n)
M <- wheat.X[,1:10]
Phenotype[test] <- NA
sERRBLUP <- sERRBLUP(M, Phenotype, 10, cores=15)

}
