% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfoldcorossvalidation.R
\name{sERRBLUP_proportions_test}
\alias{sERRBLUP_proportions_test}
\title{Cross validation for sERRBLUP}
\usage{
sERRBLUP_proportions_test(M, Pheno, k, cores = 1)
}
\arguments{
\item{M}{A {-1, 0, 1} or {0, 1} or character coded marker matrix}

\item{Pheno}{A numeric vector of phenotypes}

\item{k}{Desired proportion set of SNP interactions to be included in the model wich is proposed to be (90, 80, 70, 60, 50, 40, 30, 20, 10, 5, 1, 0.1)}

\item{cores}{The number of cores with the default value of 1}
}
\value{
A data frame of 4 components:

\describe{
  \item{Desired.Proportion}{The proportion of SNP interactions which maintained in sERRBLUP model}
  \item{PA.Effcet}{sERRBLUP predicitve ability based on effect sizes selection}
  \item{PA.Var}{sERRBLUP predicitve ability based on effect variances selection}
}
}
\description{
K fold cross validation Function with n replicatets to do phenotype prediction based sERRBLUP for finding the optimum proportion of interactions leading to highest predictive ability
}
\examples{
library(BGLR)
data(wheat)
N <- length(Phenotype)
n <- 60
test <- sample(1:N,n)
M <- wheat.X[,1:10]
Phenotype[test] <- NA
K=c(90, 80, 70, 60, 50, 40, 30, 20, 10, 5, 1, 0.1)
cross_val <- sERRBLUP_proportions_test(M, Phenotype, k=K, cores=15)

}
