% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfoldcorossvalidation.R
\name{sERRBLUP_Proportions_Test}
\alias{sERRBLUP_Proportions_Test}
\title{sERRBLUP Predictive Ability Test}
\usage{
sERRBLUP_Proportions_Test(M, Pheno, k, cores = 1)
}
\arguments{
\item{M}{A {-1, 0, 1} or {0, 1} or character coded marker matrix}

\item{Pheno}{A numeric vector of phenotypes}

\item{k}{Desired proportion set of SNP interactions to be included in the model wich is proposed to be (90, 80, 70, 60, 50, 40, 30, 20, 10, 5, 1, 0.1)}

\item{cores}{The number of cores with the default value of 1}
}
\value{
A data frame of 3 components:

\describe{
  \item{Desired.Proportion}{The proportion of SNP interactions which maintained in sERRBLUP model}
  \item{PA.Effcet}{sERRBLUP predictive ability based on effect sizes selection}
  \item{PA.Var}{sERRBLUP predictive ability based on effect variances selection}
}
}
\description{
Function to provide predictive abilities obtained from sERRBLUP for series of diffetent pairwise SNP interaction proportions to find the optimum proportion of interactions leading to the highest predictive ability
}
\examples{
library(BGLR)
data(wheat)
N <- length(Phenotype)
n <- 60
test <- sample(1:N,n)
M <- wheat.X
Phenotype[test] <- NA
K=c(90, 80, 70, 60, 50, 40, 30, 20, 10, 5, 1, 0.1)
cross_val <- sERRBLUP_Proportions_Test(M, Phenotype, k=K, cores=15)

}
