% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotypepredictionfortopinteractions.R
\name{sERRBLUP_Stepwise}
\alias{sERRBLUP_Stepwise}
\title{sERRBLUP Phenotype Prediction Function Relying On The Out Put Of sERRBLUP Relationship Matrix Function}
\usage{
sERRBLUP_Stepwise(Pheno, Gtop)
}
\arguments{
\item{Pheno}{A numeric vector of phenotypes}

\item{Gtop}{sERRBLUP Relationship matrix for the k percent of pairwise SNP interactions}
}
\value{
A numeric vector of both phenotype estimations of training set and phenotype predictions of test set based on sERRBLUP method
}
\description{
Function to do phenotype prediction based on the desired proportion of pairwise SNP interactions
}
\examples{
library(BGLR)
data(wheat)
N <- length(Phenotype)
n <- 60
test <- sample(1:N,n)
pheno <- Phenotype
pheno[test] <- NA
m <- Recodemarkers(wheat.X)
G_ERRBLUP <- Gall(m, cores=15)
G <- G_ERRBLUP$G
P <- G_ERRBLUP$P
Estimation <- SNP_Effect_Var(m, pheno, G, P, cores=15)
t_hat <- Estimation$Effect
sigma_hat <- Estimation$Effect.Var
k <- 10
Gtop_effect <- Gtop(m, t_hat, k, cores=15)
Gtop_var <- Gtop(m, sigma_hat, k, cores=15)
sERRBLUP_effect <- sERRBLUP_Stepwise(pheno, Gtop_effect)
sERRBLUP_var <- sERRBLUP_Stepwise(pheno, Gtop_var)

}
