% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllSNPeffectcalculation.R
\name{SNP_effect}
\alias{SNP_effect}
\title{Pairwise SNP Interaction Effects Function}
\usage{
SNP_effect(m, pheno_train, G_ERRBLUP)
}
\arguments{
\item{m}{{0,1,2} or {0,2} coded marker matrix with individuals in the rows and the markers in the columns}

\item{pheno_train}{A subset of one numeric phenotype vector as a training set with names for each phenotypic value}

\item{G_ERRBLUP}{EERRBLUP relationship matrix with row names and column names of all individuals}
}
\value{
A numeric vector of all estimated pairwise SNP interaction effects
}
\description{
Function to calculate all pairwise SNP interaction effects
}
\examples{
library(BGLR)
data(wheat)
pheno <- wheat.Y[1:100,1]
pheno_train <- pheno[1:round(4*length(pheno)/5)]
m <- Recodemarker(wheat.X[1:100,])
rownames(m) <- names(pheno)
G_ERRBLUP <- Gall(m)
t_hat <- SNP_effect(m, pheno_train, G_ERRBLUP)

}
