% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllSNPeffectvariancecalculation.R
\name{SNP_var}
\alias{SNP_var}
\title{Function To calculate all pairwise SNP interaction effects variances}
\usage{
SNP_var(M, pheno_train, SNP_effect)
}
\arguments{
\item{M}{{0,1,2} or {0,2} coding Marker matrix with individuals in the rows and the markers in the columns}

\item{pheno_train}{A subset of one phenotype vector as a training set with names for each phenotypic value}

\item{SNP_effect}{is a vector of all estimated pairwise SNP interaction effects}
}
\value{
A vector of all pairwise SNP interaction effects variances
}
\description{
Function to calculate all pairwise SNP interaction effects varinacs. miraculix package needs to be initially installed for this function.
}
\examples{
library(BGLR)
data(wheat)
pheno <- wheat.Y[1:100,1]
pheno_train <- pheno[1:round(4*length(pheno)/5)]
M <- Recodemarker(wheat.X[1:100,])
rownames(M) <- names(pheno)
G_all <- Gall(M)
u_hat <- SNP_effect(M, pheno_train, G_all)
u_hat_var <- SNP_var(M, pheno_train, u_hat)

}
